+++ src/app/api/order.ts
@@ -1,27 +1,6 @@
 import { instance } from '@/lib/axios';
 import { getStoreId } from './store';
-
-export type OrderRequest = {
-  restaurantTableNumber: number;
-  orderItems: { menuId: number; quantity: number }[];
-};
-
-export type OrderResponse = {
-  id: number;
-  restaurantTableId: number;
-  status: string;
-  totalPrice: number;
-  orderTime: string;
-  orderItems: {
-    orderItemId: number;
-    menuId: number;
-    menuName: string;
-    quantity: number;
-    itemPrice: number;
-    totalItemPrice: number;
-  }[];
-  paid: boolean;
-};
+import { OrderRequest, OrderResponse } from '@/types/order';
 
 const storeId = getStoreId();
